+++ CHANGELOG.md
@@ -1,2 +1,10 @@
-# v0.1.0-beta
+# Changelog
+## v0.2.0-beta
+Changes:
+- Reworked internal code for √2 and Fibonacci Sequence
+- Production code is now optimized by [python-minifier.com](https://python-minifier.com)
+- All other errors during execution of code are now caught as unknown errors
+- π - Gosper's series is not infinite anymore
+- If digits is less than 1, throws an invalid value error
+## v0.1.0-beta
 First version of Infinite Calculations!